Add: password support through PAM hook
Thank you to my old mentor Jason Tackaberry. Jason told me about pam
hook and generally how it works. So I thought it would be a nice and
easy way to have server based auth for this so it would feel like an
actual ssh server.

So I went looking for a PAM hook system for golang. To my suprise, I
stumbled upon one that Jason worked on. I was just googling around and
found one and he just happened to be one of the contributers. Kinda funny

Either way this uses the pam hook to auth and then set password in
context that way it can be reused for the container later.